# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Docmaster API",
        "description": "Docmaster API Documentation",
        "contact": {
            "name": "LJN",
            "email": "jordanndjock07@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification d'un utilisateur et retour d'un accès token",
                "operationId": "8b1ca12fe82a8104a6fbb70f5e149956",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom_utilisateur": {
                                        "type": "string",
                                        "example": "testuser"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "mdp": {
                                        "type": "string",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "1|gChmEahrYbZLpZMOdxCmokA0ntqqtwaXTokCkeHld7172a26"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Enregistrement d'un nouvel utilisateur",
                "operationId": "8f7237208878d2e15ab53c3639e56c03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prenom",
                                    "nom_famille",
                                    "nom_utilisateur",
                                    "email",
                                    "mdp",
                                    "tel",
                                    "date_naissance"
                                ],
                                "properties": {
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean Michel"
                                    },
                                    "nom_famille": {
                                        "type": "string",
                                        "example": "Atangana"
                                    },
                                    "nom_utilisateur": {
                                        "type": "string",
                                        "example": "jmatango"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jma@example.com"
                                    },
                                    "mdp": {
                                        "type": "string",
                                        "example": "secret123"
                                    },
                                    "tel": {
                                        "type": "string",
                                        "example": "+237690112233"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-21"
                                    },
                                    "localisation": {
                                        "type": "string",
                                        "example": "Yaoundé"
                                    },
                                    "infos_paiement": {
                                        "type": "string",
                                        "example": "Mobile Money"
                                    },
                                    "photo_url": {
                                        "description": "Fichier image (optionnel)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "1|gChmEahrYbZLpZMOdxCmokA0ntqqtwaXTokCkeHld7172a26"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de validation incorrectes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Les données fournies sont invalides."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne lors de la création du compte."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion de l'utilisateur",
                "description": "Nécessite un token Bearer dans l'en-tête Authorization",
                "operationId": "27275bf89a63f0655982adc39f1b8535",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnexion réussie !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "Représentation d'un utilisateur",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "nom_famille": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Joseph Johnson"
                    },
                    "nom_utilisateur": {
                        "type": "string",
                        "example": "testuser"
                    },
                    "initial_2_prenom": {
                        "type": "string",
                        "example": "JJ"
                    },
                    "tel": {
                        "type": "string",
                        "example": "237690098998"
                    },
                    "date_naissance": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01"
                    },
                    "infos_paiement": {
                        "type": "string",
                        "example": "Test payment info"
                    },
                    "localisation": {
                        "type": "string",
                        "example": "Test location"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "code_invitation": {
                        "type": "string",
                        "example": "TESTCODE"
                    },
                    "solde": {
                        "type": "number",
                        "example": 100
                    },
                    "supprime": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-30T08:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-30T08:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}